refactor: MainController 패키지 수정
